# docs/swagger.yaml
definitions:
...
  theater-ticket-system_internal_models_responses.Performance:
    properties:
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      play:
        $ref: '#/definitions/theater-ticket-system_internal_models_responses.Play'
      status:
        enum:
        - scheduled
        - completed
        - cancelled
        type: string
      updated_at:
        type: string
    required:
    - created_at
    - date
    - id
    - updated_at
    type: object
  theater-ticket-system_internal_models_responses.PerformanceSeat:
    properties:
      id:
        type: string
      performance_id:
        type: string
      price:
        type: integer
      seat:
        $ref: '#/definitions/theater-ticket-system_internal_models_responses.Seat'
      seat_id:
        type: string
      status:
        description: available, reserved, sold
        type: string
    required:
    - id
    - performance_id
    - price
    - seat_id
    - status
    type: object
  theater-ticket-system_internal_models_responses.Play:
    properties:
      author:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      genre:
        type: string
      id:
        type: string
      performances:
        items:
          $ref: '#/definitions/theater-ticket-system_internal_models_responses.Performance'
        type: array
      poster_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - author
    - created_at
    - description
    - duration
    - genre
    - id
    - poster_url
    - title
    - updated_at
    type: object
  theater-ticket-system_internal_models_responses.Seat:
    properties:
      category:
        description: parterre, balcony, box
        type: string
      hall_id:
        type: string
      id:
        type: string
      number:
        type: integer
      row:
        type: integer
    required:
    - category
    - hall_id
    - id
    - number
    - row
    type: object
info:
  contact: {}
paths:
...
  /api/halls/{id}/seats:
    get:
      description: Get static hall layout (not tied to a specific performance)
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/theater-ticket-system_internal_models_responses.Seat'
            type: array
      summary: Get hall seats
      tags:
      - seats
  /api/performances:
    get:
      description: Get list of all performances with optional filters
      parameters:
      - description: Filter by play ID
        in: query
        name: play_id
        type: string
      - description: Filter by date from (RFC3339)
        in: query
        name: date_from
        type: string
      - description: Filter by date to (RFC3339)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/theater-ticket-system_internal_models_responses.Performance'
            type: array
      summary: Get all performances
      tags:
      - performances
  /api/performances/{id}:
    get:
      description: Get detailed information about a performance
      parameters:
      - description: Performance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/theater-ticket-system_internal_models_responses.Performance'
      summary: Get performance by ID
      tags:
      - performances
  /api/performances/{id}/seats:
    get:
      description: Get hall layout with seat availability for a performance
      parameters:
      - description: Performance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/theater-ticket-system_internal_models_responses.PerformanceSeat'
            type: array
      summary: Get performance seats
      tags:
      - performances
...
swagger: "2.0"